Generate %changelog blocks differently and fix Source0
This is a result of the package review process.  The %changelog
generation process I was doing was similar to what we had back when I
started working on anaconda in 2005 which was we used the %changelog
as the upstream changelog for the project.  Anaconda did not make
releases that were not an SRPM in the distribution.  Personally, I
like seeing a summary like that in the RPM %changelog, but I am also
entirely ok with there being no RPM %changelog.  We're not quite there
yet in Fedora and related distributions, so the prevailing style seems
to be to reserve the RPM %changelog for changes to the package itself
and not the upstream project.

Since this project is set up to cut a new release more or less for
every change, that means the downstream packaging effort is not where
work happens.  Also, the changelog entries from new releases will just
be of the form "- Upgrade to rpminspect-data-centos-1.0" and nothing
else.  However, that does not stop downstream distributions from
rebuilding and shifting things around and otherwise doing stuff to
packages.  That's fine, but there is a desire to retain those
changelog entries from downstream.  But I don't want to commit them
upstream and deal with this manual syncing process.  So here's what
I've done:

1) Removed %changelog from the spec file template.

2) Removed the Source1 entry so there is no longer a 'changelog' file
that is %include'd.

3) Removed mkrpmchangelog.sh

4) For Copr builds, the .copr/Makefile will generate the same
copr-style changelog I was doing which is just a single entry.  It
will tack that on to the generated spec file and away it goes.

5) For released builds, the submit-koji-build.sh script will collect
the downstream changelog and then prepend a new entry of the style
"upgrade to ..." on the top.  It will then tack that on to the end of
the generated spec file after the %changelog line.  This will be
committed downstream.

If all goes well, this should preserve downstream changelog
modifications while getting the upstream project out of the business
of caring about it manually.